requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.1
django-cache-url==3.4.5
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
environs==10.3.0
marshmallow==3.20.2
packaging==23.2
psycopg==3.1.17
psycopg-binary==3.1.17
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
